src/pyproject.toml
[tool.poetry]
name = "WebChat"
version = "0.1.0"
description = ""
authors = []
readme = "../README.md"


[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.109.2"}
fastapi-keycloak-middleware = "^0.3.1"
pydantic = {extras = ["dotenv"], version = "^2.6.1"}
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"
loguru = "^0.7.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
ruff = "^0.2.2"
mypy = "^1.8.0"


[tool.ruff]
target-version = "py312"
line-length = 120


[tool.ruff.lint]
extend-select = ["I"]


[tool.mypy]
allow_redefinition = "False"
check_untyped_defs = "True"
disallow_untyped_decorators = "True"
disallow_any_explicit = "False"
disallow_any_generics = "True"
disallow_untyped_calls = "True"
ignore_errors = "False"
ignore_missing_imports = "True"
implicit_reexport = "False"
local_partial_types = "True"
strict_optional = "True"
strict_equality = "True"
no_implicit_optional = "True"
warn_unused_ignores = "True"
warn_redundant_casts = "True"
warn_unused_configs = "True"
warn_unreachable = "True"
warn_no_return = "True"
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
exclude = ["migrations"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
